removed warnings on windows simplified Makevars file
git-svn-id: http://dyncall.org/svn/dyncall/trunk@837 c71cb566-5a31-4692-9165-cf76bdc42cab

+++ bindings/R/rdyncall/src/rutils.c
@@ -47,8 +47,8 @@ SEXP r_asextptr(SEXP x)
   switch(TYPEOF(x))
   {
     case NILSXP: addr = NULL; break;
-    case INTSXP: addr = (void*) (unsigned long) INTEGER(x)[0]; break;
-    case REALSXP: addr = (void*) (unsigned long) REAL(x)[0]; break;
+    case INTSXP: addr = (void*) (ptrdiff_t) INTEGER(x)[0]; break;
+    case REALSXP: addr = (void*) (ptrdiff_t) REAL(x)[0]; break;
     default: error("invalid type");
   }
   return R_MakeExternalPtr( addr, R_NilValue, R_NilValue );
